+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+- 0.3.7
+  - fix: filter mode with empty files
+  - fix: action class to poo mode
 - 0.3.6
   - add compact mode to run
 - 0.3.5: